Minor changes to avoid compiler warnings
git-svn-id: https://subversion.xor.aps.anl.gov/synApps/areaDetector/trunk@15497 dc6c5ff5-0b8b-c028-a01f-ffb33f00fc8b

+++ ADApp/pluginSrc/NDPluginProcess.h
@@ -127,12 +127,12 @@ protected:
                                 
 private:
     NDArray *pBackground;
-    int     nBackgroundElements;
+    size_t  nBackgroundElements;
     NDArray *pFlatField;
-    int     nFlatFieldElements;
+    size_t  nFlatFieldElements;
     NDArray *pFilter;
-    int     numFiltered;
+    int  numFiltered;
 };
-#define NUM_NDPLUGIN_PROCESS_PARAMS (&LAST_NDPLUGIN_PROCESS_PARAM - &FIRST_NDPLUGIN_PROCESS_PARAM + 1)
+#define NUM_NDPLUGIN_PROCESS_PARAMS ((int)(&LAST_NDPLUGIN_PROCESS_PARAM - &FIRST_NDPLUGIN_PROCESS_PARAM + 1))
     
 #endif
